캐릭터 HP_MP 확인 가능

+++ second-poject/Assets/01. Scripts/Skill/SO_Skill.cs
@@ -9,8 +9,8 @@ public class SO_Skill : ScriptableObject
     public Elements skillElements;
     public float skillDamage;
     public float needFOC;
-    public float needMp;
-    public float needHp;
+    public int needMp;
+    public int needHp;
 
     public bool categorPyhysics;
     public bool categoryChemistry;
